Support options.parse and make options.ast Boolean.
It's vitally important that the parsed AST corresponds to the given
source code, so it's better for Reify to do its own parsing, rather than
accepting a pre-parsed AST.

It's also important to use the optional options.parse function when
parsing code generated by the compiler, since different parsers generate
different AST node types, even for the tiny subset of the language
needed by Reify's generated code. In particular, string literals are
"StringLiteral" nodes according to Babylon, but string-valued "Literal"
nodes according to most other parsers.

If options.ast is truthy, then the compile function will return the
modified AST as the .ast property of its result object. Note that the
compile function previously returned a string rather than an object. The
equivalent string is now always returned as result.code. This API is
similar to the one provided by Babel, for what it's worth.

This commit undoes some of the work of other recent commits, since
options.ast was previously a way of providing a pre-parsed AST, and that
use case is no longer supported. There was never a published version of
Reify that honored options.ast in that way, so I don't feel too bad
about changing the meaning of the option.