Prevent potential setter leaks for module.import calls in loops.
By statically identifying each module.import call with a unique key, we
can dispose of any previous setter functions registered with the same key
by the same parent module.

This is a backwards-compatible change, since the key argument is optional,
and the third argument to module.import was previously ignored.