Standardize format of specifier maps.
Instead of inverting the specifier map depending on whether it's
representing imports or exports, the keys are now always names that were
exported by some module, be it the current module (because we're exporting
them) or some other module (because we're importing them). The values of
the specifier map are arrays of local variable names that are bound to
those exported/imported names.

This representation fixes a bug that mistakenly prevented

  import { a as x, a as y } from "./module";

from working, because we now map "a" to ["x", "y"], whereas previously
we mapped "x" to "a" and "y" to "a", but we weren't doing anything to
detect that duplication.