Support compileOptions.avoidArrowFunctions.
When truthy, this option allows disabling arrow functions in generated
code. We currently only generate arrow functions for module.export calls,
because creating arrow functions has overhead in Node 4, but those arrow
functions can be problematic in older environments.

This commit should be totally backwards compatible because the default
value for this option is false.