Use the same entryKey across any number of Reify copies.
When we're able to use a Symbol, the entryKey is always the same, as it
should be, even if multiple copies of Reify are interacting.

Before this commit, when Symbol was not available, we used makeUniqueKey()
to generate a unique entryKey every time, which seems like a pretty
undesirable difference in behavior from the Symbol case.