Use module.watch(require(id), ...) to avoid calling module.require.
If we choose to go in this direction, we would be removing a significant
barrier to interop with CommonJS module systems other than Node (in
particular, Webpack and Browserify), since module.require would no longer
be a requirement, and third-party bundling tools could easily understand
dependencies embedded in module.watch(require(id), ...) calls.

I'm not committed to module.watch as a method name, though it's short and
accurately reflects that the parent module is watching an exports object
returned by a child module.