Add failing test due to missing module.runSetters() call.
This test fails because module.runSetters() is not called when the
test/setter/cycle/cjs.js module finishes evaluating, due to this commit:
7932aa56cef2ead61084c3b67a5afa329fe5415e

The general problem is that we need to call module.runSetters() even for
modules that were not compiled by Reify, or wrapped with module.run(...),
so that other modules can import live bindings from them.

If you add module.runSetters() to the end of cjs.js, or add back the
module.runSetters() call in the compile hook, this test will pass.